requirements.txt
discord.py-self>=2.0.0
PyNaCl>=1.5.0
